made new enemy bullet

+++ Assets/Scripts/Weapons/Weapon.cs
@@ -85,7 +85,7 @@ public class Weapon : MonoBehaviour
         Bullet bulletScript = bullet.GetComponent<Bullet>();
         if (bulletScript != null)
         {
-            bulletScript.damage = damage;
+            bulletScript.damage = (int)damage;
         }
         
         animator.SetTrigger("Shoot");
